+++ src/release/mod.rs
@@ -17,13 +17,11 @@ It supports both the [v1] and [v2] specs.
 
 */
 
-use crate::dist::*;
-use crate::util;
-
+use crate::{dist::*, error::Error, util};
 use chrono::{DateTime, Utc};
 use serde::{de, Deserialize, Deserializer, Serialize};
 use serde_json::Value;
-use std::{borrow::Borrow, collections::HashMap, error::Error, fs::File, path::Path};
+use std::{borrow::Borrow, collections::HashMap, fs::File, path::Path};
 
 mod v1;
 mod v2;
@@ -160,11 +158,11 @@ impl Release {
 
     /// Deserializes `meta`, which contains PGXN distribution release
     /// metadata, into a [`Release`].
-    fn from_version(version: u8, meta: Value) -> Result<Self, Box<dyn Error>> {
+    fn from_version(version: u8, meta: Value) -> Result<Self, Error> {
         match version {
             1 => v1::from_value(meta),
             2 => v2::from_value(meta),
-            _ => Err(Box::from(format!("Unknown meta version {version}"))),
+            _ => Err(Error::UnknownSpec),
         }
 
         // XXX: Add signature validation.
@@ -173,7 +171,7 @@ impl Release {
     /// Loads the release `META.json` data from `file` then converts into a
     /// [`Release`]. Returns an error on file error or if the content of
     /// `file` is not valid PGXN `META.json` data.
-    pub fn load<P: AsRef<Path>>(file: P) -> Result<Self, Box<dyn Error>> {
+    pub fn load<P: AsRef<Path>>(file: P) -> Result<Self, Error> {
         let meta: Value = serde_json::from_reader(File::open(file)?)?;
         meta.try_into()
     }
@@ -266,14 +264,13 @@ impl Release {
 }
 
 impl TryFrom<Value> for Release {
-    type Error = Box<dyn Error>;
+    type Error = Error;
     /// Converts the PGXN release `META.json` data from `meta` into a
     /// [`Release`]. Returns an error if `meta` is invalid.
     ///
     /// # Example
     ///
     /// ``` rust
-    /// # use std::error::Error;
     /// use serde_json::json;
     /// use pgxn_meta::release::*;
     ///
@@ -309,16 +306,13 @@ impl TryFrom<Value> for Release {
     fn try_from(meta: Value) -> Result<Self, Self::Error> {
         // Make sure it's valid.
         let mut validator = crate::valid::Validator::new();
-        let version = match validator.validate_release(&meta) {
-            Err(e) => return Err(Box::from(e.to_string())),
-            Ok(v) => v,
-        };
+        let version = validator.validate_release(&meta)?;
         Release::from_version(version, meta)
     }
 }
 
 impl TryFrom<&[&Value]> for Release {
-    type Error = Box<dyn Error>;
+    type Error = Error;
     /// Merge multiple PGXN release `META.json` data from `meta` into a
     /// [`Release`]. Returns an error if `meta` is invalid.
     ///
@@ -329,7 +323,6 @@ impl TryFrom<&[&Value]> for Release {
     /// # Example
     ///
     /// ``` rust
-    /// # use std::error::Error;
     /// use serde_json::json;
     /// use pgxn_meta::release::*;
     ///
@@ -369,12 +362,11 @@ impl TryFrom<&[&Value]> for Release {
     /// [RFC 7396]: https:///www.rfc-editor.org/rfc/rfc7396.html
     fn try_from(meta: &[&Value]) -> Result<Self, Self::Error> {
         if meta.is_empty() {
-            return Err(Box::from("meta contains no values"));
+            return Err(Error::Param("meta contains no values"));
         }
 
         // Find the version of the first doc.
-        let version =
-            util::get_version(meta[0]).ok_or("No spec version found in first meta value")?;
+        let version = util::get_version(meta[0]).ok_or_else(|| Error::UnknownSpec)?;
 
         // Convert the first doc to v2 if necessary.
         let mut v2 = match version {
@@ -390,21 +382,20 @@ impl TryFrom<&[&Value]> for Release {
 
         // Validate the patched doc and return.
         let mut validator = crate::valid::Validator::new();
-        validator.validate_release(&v2).map_err(|e| e.to_string())?;
+        validator.validate_release(&v2)?;
         Release::from_version(2, v2)
     }
 }
 
 impl TryFrom<Release> for Value {
-    type Error = Box<dyn Error>;
+    type Error = Error;
     /// Converts PGXN release `meta` into a [serde_json::Value].
     ///
     /// # Example
     ///
     /// ``` rust
-    /// # use std::error::Error;
     /// use serde_json::{json, Value};
-    /// use pgxn_meta::release::*;
+    /// use pgxn_meta::{error::Error, release::*};
     ///
     /// let meta_json = json!({
     ///   "name": "pair",
@@ -434,7 +425,7 @@ impl TryFrom<Release> for Value {
     ///
     /// let meta = Release::try_from(meta_json);
     /// assert!(meta.is_ok());
-    /// let val: Result<Value, Box<dyn Error>> = meta.unwrap().try_into();
+    /// let val: Result<Value, Error> = meta.unwrap().try_into();
     /// assert!(val.is_ok());
     /// ```
     fn try_from(meta: Release) -> Result<Self, Self::Error> {
@@ -444,7 +435,7 @@ impl TryFrom<Release> for Value {
 }
 
 impl TryFrom<&String> for Release {
-    type Error = Box<dyn Error>;
+    type Error = Error;
     /// Converts `str` into JSON and then into a [`Release`]. Returns an
     /// error if the content of `str` is not valid PGXN `META.json` data.
     fn try_from(str: &String) -> Result<Self, Self::Error> {
@@ -454,7 +445,7 @@ impl TryFrom<&String> for Release {
 }
 
 impl TryFrom<Release> for String {
-    type Error = Box<dyn Error>;
+    type Error = Error;
     /// Converts `meta` into a JSON String.
     fn try_from(meta: Release) -> Result<Self, Self::Error> {
         let val = serde_json::to_string(&meta)?;